+++ src/index.html
@@ -23,14 +23,17 @@
       </ul>
     </div>
   </header>
-  <section class="screen" id="home">
-    <img class="home-img" src="homepokemon.jpg">
-    <main class="home-content">
-      <h1>Lorem</h1>
-      <p>Lorem, ipsum dolor sit amet consectetur adipisicing elit. Illum, natus magni. 
-        Temporibus nostrum quibusdam harum iusto aperiam aut, sit, possimus tenetur minus voluptate 
-        deserunt nihil rerum, reiciendis hic autem nemo!</p>
-    </main>
+  <section class="screen flex" id="home">
+    <!-- <h1 class="header-title">HOME</h1> -->
+    <div class="home-box">
+      <div class="home-content">
+        <h2>SOBRE</h2>
+        <p>Olá, treinador pokémon! Você quer filtrar seus pokémons por tipos, fraquezas e altura? 
+          Quer calcular quantos candys o seu pokémon precisa para evoluir e qual é o CP resultante da evolução?
+          Quer visualizar porcentagens de pokémons por tipo? Então esta aplicação é ideal para você!
+        </p>
+      </div>
+    </div>
   </section>
   <section class="screen" id="ranking">
     <h1 class="header-title">ESTATÍSTICAS</h1>
