+++ src/style.css
@@ -277,14 +277,28 @@ main{
     width: 100vh;
     height: auto;
 }
-.home-img {
-    width: 207vh;
+
+.flex {
+    display: flex;
+    align-items: center;
+    justify-content: center;
+}
+
+.home-box {
+    /* width: 100%;
+    height: 100%; */
+    margin: 3px;
+    padding: 2%;
+    width: 100vh;
     height: auto;
+    background-color: teal;
 }
 
 .home-content {
     background-color: white;
-    justify-content: center;
+    position: absolute;
+    text-align: center;
+    align-self: center;
 }
 .filter-config{  
     margin: 5px;
